[Services] Fix Clang errors with tunnelbroker
Summary:
Clang doesn't provide an alias for usize_t which
corresponds directly to amqp's overloaded types. So
explicity cast to a type which amqp has an overload for.

Test Plan:
Follow test plan from later tunnelbroker diff with
MacOS cmake fixes

Reviewers: atul, max, karol

Subscribers: ashoat, tomek, adrian, abosh

Differential Revision: https://phab.comm.dev/D4765

+++ services/tunnelbroker/src/Amqp/AmqpManager.cpp
@@ -40,8 +40,8 @@ void AmqpManager::connectInternal() {
   });
 
   AMQP::Table arguments;
-  arguments["x-message-ttl"] = AMQP_MESSAGE_TTL;
-  arguments["x-expires"] = AMQP_QUEUE_TTL;
+  arguments["x-message-ttl"] = (uint64_t)AMQP_MESSAGE_TTL;
+  arguments["x-expires"] = (uint64_t)AMQP_QUEUE_TTL;
   this->amqpChannel->declareExchange(fanoutExchangeName, AMQP::fanout);
   this->amqpChannel->declareQueue(tunnelbrokerID, AMQP::durable, arguments)
       .onSuccess([this, tunnelbrokerID, fanoutExchangeName](
